requirements.txt
# requirements.txt
langgraph
langchain-google-genai
pandas
pdfplumber
pytest
google-generativeai